Refactor tests to use stubs
