+++ tests/DataCacheTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace {
     use PHPUnit\Framework\TestCase;
     use Lotgd\DataCache;
+    use Lotgd\Tests\Stubs\CacheDummySettings;
 
     require_once __DIR__ . '/../config/constants.php';
 
@@ -12,38 +13,6 @@ namespace {
         define('DATACACHE_FILENAME_PREFIX', 'datacache-');
     }
 
-    if (!class_exists('CacheDummySettings')) {
-        class CacheDummySettings
-        {
-            private array $values;
-
-            public function __construct(array $values = [])
-            {
-                $this->values = $values;
-            }
-
-            public function getSetting(string|int $name, mixed $default = false): mixed
-            {
-                return $this->values[$name] ?? $default;
-            }
-
-            public function loadSettings(): void
-            {
-            }
-            public function clearSettings(): void
-            {
-            }
-            public function saveSetting(string|int $name, mixed $value): bool
-            {
-                $this->values[$name] = $value;
-                return true;
-            }
-            public function getArray(): array
-            {
-                return $this->values;
-            }
-        }
-    }
 
     final class DataCacheTest extends TestCase
     {
