+++ tests/DumpOutputTest.php
@@ -18,41 +18,12 @@ namespace Lotgd {
 namespace {
     use PHPUnit\Framework\TestCase;
     use Lotgd\OutputArray;
+    use Lotgd\Tests\Stubs\DumpDummySettings;
     use Lotgd\DumpItem;
-    use Lotgd\Settings;
 
     require_once __DIR__ . '/../config/constants.php';
     require_once __DIR__ . '/../lib/settings.php';
 
-    if (!class_exists('DumpDummySettings')) {
-        class DumpDummySettings extends Settings
-        {
-            private array $values;
-            public function __construct(array $values = [])
-            {
-                $this->values = $values;
-            }
-            public function getSetting(string|int $name, mixed $default = false): mixed
-            {
-                return $this->values[$name] ?? $default;
-            }
-            public function loadSettings(): void
-            {
-            }
-            public function clearSettings(): void
-            {
-            }
-            public function saveSetting(string|int $name, mixed $value): bool
-            {
-                $this->values[$name] = $value;
-                return true;
-            }
-            public function getArray(): array
-            {
-                return $this->values;
-            }
-        }
-    }
 
     final class DumpOutputTest extends TestCase
     {
