+++ tests/MailTest.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace {
     use PHPUnit\Framework\TestCase;
+    use Lotgd\Tests\Stubs\Database;
+    use Lotgd\Tests\Stubs\PHPMailer;
+    use Lotgd\Tests\Stubs\MailDummySettings;
     use Lotgd\Mail;
 
     require_once __DIR__ . '/../config/constants.php';
@@ -23,130 +26,6 @@ namespace {
     'notificationmailsubject' => '{subject}',
     'notificationmailtext' => '{body}',
     ];
-}
-
-// --- Database stub ---
-namespace Lotgd\MySQL {
-    if (!class_exists('Lotgd\\MySQL\\Database', false)) {
-        class Database
-        {
-            public static array $settings_table = [];
-            public static int $onlineCounter = 0;
-            public static int $affected_rows = 0;
-
-            public static function prefix(string $name, bool $force = false): string
-            {
-                return $name;
-            }
-
-            public static function query(string $sql, bool $die = true)
-            {
-                global $accounts_table, $mail_table, $last_query_result;
-                if (preg_match("/SELECT prefs,emailaddress FROM accounts WHERE acctid='?(\d+)'?;/", $sql, $m)) {
-                    $acctid = (int)$m[1];
-                    $row = $accounts_table[$acctid] ?? ['prefs' => '', 'emailaddress' => ''];
-                    $last_query_result = [$row];
-                    return $last_query_result;
-                }
-                if (strpos($sql, 'INSERT INTO mail') === 0) {
-                    if (preg_match("/\((?:'|\")?(\d+)(?:'|\")?,(?:'|\")?(\d+)(?:'|\")?,(?:'|\")?(.*?)(?:'|\")?,(?:'|\")?(.*?)(?:'|\")?,(?:'|\")?(.*?)(?:'|\")?\)/", $sql, $m)) {
-                            $from = (int)$m[1];
-                            $to = (int)$m[2];
-                            $subject = $m[3];
-                            $body = $m[4];
-                            $sent = $m[5];
-                    } else {
-                        $from = $to = 0;
-                        $subject = '';
-                        $body = '';
-                        $sent = '';
-                    }
-                    $id = count($mail_table) + 1;
-                    $mail_table[] = ['messageid' => $id,'msgfrom' => $from,'msgto' => $to,'subject' => $subject,'body' => $body,'sent' => $sent,'seen' => 0];
-                    $last_query_result = true;
-                    return true;
-                }
-                if (preg_match("/SELECT name FROM accounts WHERE acctid='?(\d+)'?;/", $sql, $m)) {
-                    $acctid = (int)$m[1];
-                    $row = ['name' => $accounts_table[$acctid]['name'] ?? ''];
-                    $last_query_result = [$row];
-                    return $last_query_result;
-                }
-                if (preg_match("/SELECT count\(messageid\) AS count FROM mail WHERE msgto=(\d+)(.*)/", $sql, $m)) {
-                    $userId = (int)$m[1];
-                    $onlyUnread = strpos($sql, 'seen=0') !== false;
-                    $count = 0;
-                    foreach ($mail_table as $row) {
-                        if ($row['msgto'] == $userId && (!$onlyUnread || $row['seen'] == 0)) {
-                            $count++;
-                        }
-                    }
-                    $last_query_result = [[ 'count' => $count ]];
-                    return $last_query_result;
-                }
-                if (strpos($sql, 'SELECT count(acctid) as counter FROM accounts') === 0) {
-                    $last_query_result = [[ 'counter' => self::$onlineCounter ]];
-                    return $last_query_result;
-                }
-                if (preg_match('/SELECT \* FROM (.+)/', $sql, $m)) {
-                    if ($m[1] === 'settings') {
-                        $last_query_result = [];
-                        foreach (self::$settings_table as $k => $v) {
-                            $last_query_result[] = ['setting' => $k,'value' => $v];
-                        }
-                        return $last_query_result;
-                    }
-                }
-                if (strpos($sql, 'INSERT INTO settings') === 0) {
-                    if (preg_match('/VALUES\((.+),(.+)\)/', $sql, $m)) {
-                        $name = trim($m[1], "'\"");
-                        $value = trim($m[2], "'\"");
-                    } else {
-                        $name = $value = '';
-                    }
-                    self::$settings_table[$name] = $value;
-                    self::$affected_rows = 1;
-                    $last_query_result = true;
-                    return true;
-                }
-                if (preg_match('/UPDATE (.+) SET value=(.+) WHERE setting=(.+)/', $sql, $m)) {
-                    if ($m[1] === 'settings') {
-                        $value = trim($m[2], "'\"");
-                        $name = trim($m[3], "'\"");
-                        if (isset(self::$settings_table[$name])) {
-                            self::$settings_table[$name] = $value;
-                            self::$affected_rows = 1;
-                        } else {
-                            self::$affected_rows = 0;
-                        }
-                        $last_query_result = true;
-                        return true;
-                    }
-                }
-                $last_query_result = [];
-                return [];
-            }
-            public static function fetchAssoc(array|\mysqli_result &$result)
-            {
-                return array_shift($result);
-            }
-
-            public static function freeResult(array|\mysqli_result &$result): bool
-            {
-                $result = null;
-                return true;
-            }
-
-            public static function numRows(array|\mysqli_result $result): int
-            {
-                return is_array($result) ? count($result) : 0;
-            }
-            public static function affectedRows(): int
-            {
-                return self::$affected_rows;
-            }
-        }
-    }
 }
 
 namespace {
@@ -185,87 +64,14 @@ namespace {
         }
     }
 }
-// --- Class stubs ---
-
-namespace PHPMailer\PHPMailer {
-    class PHPMailer
-    {
-        public array $to = [];
-        public array $cc = [];
-        public array $reply = [];
-        public $Body = '';
-        public $AltBody = '';
-        public $Subject = '';
-        public function __construct($exc = false)
-        {
-        }
-        public function IsSendmail()
-        {
-        }
-        public function isSMTP()
-        {
-        }
-        public function AddReplyTo($addr, $name = '')
-        {
-            $this->reply[$addr] = $name;
-        }
-        public function AddAddress($addr, $name = '')
-        {
-            $this->to[$addr] = $name;
-        }
-        public function AddCC($addr, $name = '')
-        {
-            $this->cc[$addr] = $name;
-        }
-        public function SetLanguage($lang)
-        {
-        }
-        public function IsHTML($v = true)
-        {
-        }
-        public function Send()
-        {
-            $GLOBALS['mail_sent_count'] = ($GLOBALS['mail_sent_count'] ?? 0) + 1;
-        }
-    }
-}
 
 namespace {
     use PHPUnit\Framework\TestCase;
     use Lotgd\Mail;
+    use Lotgd\Tests\Stubs\Database;
+    use Lotgd\Tests\Stubs\PHPMailer;
+    use Lotgd\Tests\Stubs\MailDummySettings;
 
-    if (!class_exists('MailDummySettings')) {
-        class MailDummySettings
-        {
-            private array $values;
-
-            public function __construct(array $values = [])
-            {
-                $this->values = $values;
-            }
-
-            public function getSetting(string|int $settingname, mixed $default = false): mixed
-            {
-                return $this->values[$settingname] ?? $default;
-            }
-
-            public function loadSettings(): void
-            {
-            }
-            public function clearSettings(): void
-            {
-            }
-            public function saveSetting(string|int $settingname, mixed $value): bool
-            {
-                $this->values[$settingname] = $value;
-                return true;
-            }
-            public function getArray(): array
-            {
-                return $this->values;
-            }
-        }
-    }
 
     final class MailTest extends TestCase
     {
